feat(#2147): move move FuncXslTest under ParsingTrainTest
